arm: vc1sdp: Change stride parameters to ptrdiff_t
This was missed in db54426975e124e98e5130ad01316cb7afd60630.

+++ libavcodec/arm/vc1dsp_init_neon.c
@@ -33,12 +33,12 @@ void ff_vc1_inv_trans_4x8_dc_neon(uint8_t *dest, ptrdiff_t stride, int16_t *bloc
 void ff_vc1_inv_trans_8x4_dc_neon(uint8_t *dest, ptrdiff_t stride, int16_t *block);
 void ff_vc1_inv_trans_4x4_dc_neon(uint8_t *dest, ptrdiff_t stride, int16_t *block);
 
-void ff_vc1_v_loop_filter4_neon(uint8_t *src, int stride, int pq);
-void ff_vc1_h_loop_filter4_neon(uint8_t *src, int stride, int pq);
-void ff_vc1_v_loop_filter8_neon(uint8_t *src, int stride, int pq);
-void ff_vc1_h_loop_filter8_neon(uint8_t *src, int stride, int pq);
-void ff_vc1_v_loop_filter16_neon(uint8_t *src, int stride, int pq);
-void ff_vc1_h_loop_filter16_neon(uint8_t *src, int stride, int pq);
+void ff_vc1_v_loop_filter4_neon(uint8_t *src, ptrdiff_t stride, int pq);
+void ff_vc1_h_loop_filter4_neon(uint8_t *src, ptrdiff_t stride, int pq);
+void ff_vc1_v_loop_filter8_neon(uint8_t *src, ptrdiff_t stride, int pq);
+void ff_vc1_h_loop_filter8_neon(uint8_t *src, ptrdiff_t stride, int pq);
+void ff_vc1_v_loop_filter16_neon(uint8_t *src, ptrdiff_t stride, int pq);
+void ff_vc1_h_loop_filter16_neon(uint8_t *src, ptrdiff_t stride, int pq);
 
 void ff_put_pixels8x8_neon(uint8_t *block, const uint8_t *pixels,
                            ptrdiff_t line_size, int rnd);
